move some functions from the ui.js to the respective controllers, stop requiring models into ui
